+++ test/unit/response.test.js
@@ -1,6 +1,7 @@
 var fs = require('fs'),
     _ = require('lodash'),
     expect = require('expect.js'),
+    reason = require('http-reasons'),
     util = require('../../lib/util'),
     request = require('postman-request'),
 
@@ -11,6 +12,16 @@ var fs = require('fs'),
 
 /* global describe, it */
 describe('Response', function () {
+    describe('isResponse', function () {
+        it('must distinguish between responses and other objects', function () {
+            var response = new Response(),
+                nonResponse = {};
+
+            expect(Response.isResponse(response)).to.be(true);
+            expect(Response.isResponse(nonResponse)).to.be(false);
+        });
+    });
+
     describe('json representation', function () {
         it('must match what the response was initialized with', function () {
             var rawResponse = fixtures.collectionV2.item[0].response[0],
@@ -49,6 +60,59 @@ describe('Response', function () {
 
             // Skip cookie tests, because cookies are tested independently.
             expect(jsonified).to.have.property('cookie');
+        });
+    });
+
+    describe('details', function () {
+        it('should correctly accept provided details', function () {
+            expect(new Response({ code: 200 }).details()).to.eql({
+                name: '200 OK',
+                detail: reason.lookup(200).detail,
+                code: 200
+            });
+        });
+
+        it('should correctly set a flag for server reasons', function () {
+            expect(new Response({ code: 200, _reasonFromServer: true }).details()).to.eql({
+                name: '200 OK',
+                detail: reason.lookup(200).detail,
+                code: 200,
+                fromServer: true
+            });
+        });
+
+        it('should correctly update the reason and response code where applicable', function () {
+            var response = new Response({ code: 200 });
+
+            expect(response.details()).to.eql({
+                name: '200 OK',
+                detail: reason.lookup(200).detail,
+                code: 200,
+                fromServer: true
+            });
+
+            response.update({ code: 201, _reasonFromServer: true });
+
+            expect(response.details()).to.eql({
+                name: 'Created',
+                detail: reason.lookup('201').detail,
+                code: 201,
+                fromServer: true
+            });
+        });
+
+        it('should correctly check for response code disparities', function () {
+            var response = new Response();
+
+            // hijacked update
+            response.code = 201;
+            response.status = '201 Created';
+
+            expect(response.details()).to.eql({
+                name: '201 Created',
+                detail: reason.lookup(201).detail,
+                code: 201
+            });
         });
     });